assign/ubq3/requirements.txt
ollama
flask-cors